Refactor python project classes
