+++ groove/groove-python/groove/assets/__init__.py
@@ -1,6 +1,5 @@
-from pathlib import Path
-
 from importlib.resources import files
+from pathlib import Path
 
 
 def get_asset_path(asset_name: str) -> Path:
