+++ groove/groove-python/groove/tests/test_auth.py
@@ -6,8 +6,8 @@ from playwright._impl._api_types import Error as PlaywrightError
 from requests import get
 
 from groove.assets import get_asset_path
-from groove.proxy import (Groove, TapeRecord, TapeRequest, TapeResponse,
-                          TapeSession)
+from groove.proxy import Groove
+from groove.tape import  TapeRecord, TapeRequest, TapeResponse, TapeSession
 
 AUTH_USERNAME = "test-username"
 AUTH_PASSWORD = "test-password"
