+++ groove/proxy/go.mod
@@ -6,9 +6,9 @@ go 1.18
 
 require (
 	github.com/gin-gonic/gin v1.8.1
-	github.com/piercefreeman/goproxy v0.0.5
+	github.com/piercefreeman/goproxy v0.0.6
 	github.com/pquerna/cachecontrol v0.1.0
-	github.com/refraction-networking/utls v1.1.3
+	github.com/refraction-networking/utls v1.1.5
 	golang.org/x/net v0.0.0-20221012135044-0b7e1fb9d458
 )
 
@@ -20,14 +20,14 @@ require (
 	github.com/go-playground/validator/v10 v10.10.0 // indirect
 	github.com/goccy/go-json v0.9.7 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
-	github.com/klauspost/compress v1.13.6 // indirect
+	github.com/klauspost/compress v1.15.9 // indirect
 	github.com/leodido/go-urn v1.2.1 // indirect
 	github.com/mattn/go-isatty v0.0.14 // indirect
 	github.com/modern-go/concurrent v0.0.0-20180228061459-e0a39a4cb421 // indirect
 	github.com/modern-go/reflect2 v1.0.2 // indirect
 	github.com/pelletier/go-toml/v2 v2.0.1 // indirect
 	github.com/ugorji/go/codec v1.2.7 // indirect
-	golang.org/x/crypto v0.0.0-20211108221036-ceb1ce70b4fa // indirect
+	golang.org/x/crypto v0.0.0-20220829220503-c86fa9a7ed90 // indirect
 	golang.org/x/sys v0.0.0-20220728004956-3c1f35247d10 // indirect
 	golang.org/x/text v0.3.7 // indirect
 	google.golang.org/protobuf v1.28.0 // indirect
